+++ src/games/TragamonedasB.ts
@@ -1,31 +1,20 @@
 import { Juego } from "../models/Juego";
 import { Jugador } from "../models/Jugador";
+import { Tragamonedas } from "../models/Tragamonedas";
 import { Menu } from "../utils/Menu";
 import * as fs from "node:fs";
 
-export class DeluxeCachinEasyWin extends Juego {
-  private apuestaMinima: number;
-  private apuestaMaxima: number;
-  private simbolos: string[];
-  private jugada: string[];
-  private valoresSimbolos: Record<string, number> = {
-    "🤴": 8,
-    "🧙": 10,
-    "🦄": 4,
-    "👑": 9,
-  };
-  private ganancia: number;
-  private montoApostado: number;
-  private saldoInicial: number;
+export class DeluxeCachinEasyWin extends Tragamonedas {
   public constructor() {
     super();
     this.apuestaMinima = 50;
     this.apuestaMaxima = 750;
-    this.simbolos = ["🦄", "🧙", "🤴", "👑"];
+    this.simbolos = ["🦄", "🧙", "🤴", "👑", "🦄", "🧙"]; //Espacio de 6, pero con dos repetidos para hacer honor al "easy win".
     this.jugada = [];
     this.ganancia = 0; // inicializa en 0 porque aun no hay ganancia
-    this.montoApostado = 50;
+    this.apuesta = this.apuestaMinima;
     this.saldoInicial = 0;
+    this.nombreTragamonedas = "Deluxe Cachin Easy Win";
   }
   private readonly tiros: number = 5;
 
@@ -44,10 +33,10 @@ export class DeluxeCachinEasyWin extends Juego {
 
     // Opciones del jugador dentro del juego
     this.saldoInicial = jugador.obtenerSaldo();
-    this.interfaceCachin(jugador);
+    this.interfaceTragamonedas(jugador);
 
-    this.montoApostado = await this.pedirMonto(jugador);
-    if (this.montoApostado === 0) {
+    this.apuesta = await this.pedirApuesta(jugador);
+    if (this.apuesta === 0) {
       const confirmacion = await Menu.pedirConfirmacion(
         "Estás seguro que deseas salir?"
       );
@@ -82,13 +71,13 @@ export class DeluxeCachinEasyWin extends Juego {
     while (opcion !== "salir") {
       opcion = await Menu.elegirOpcion("¿Que deseas hacer?", opciones);
       if (opcion === "tirada") {
-        if (jugador.obtenerSaldo() < this.montoApostado) {
+        if (jugador.obtenerSaldo() < this.apuesta) {
           console.log("Tu saldo en insuficiente");
           await this.esperar(3);
         }
-        jugador.restarSaldo(this.montoApostado);
-        await this.interfaceCachin(jugador);
-        console.log(await this.tirada());
+        jugador.restarSaldo(this.apuesta);
+        await this.interfaceTragamonedas(jugador);
+        await this.tirada();
         console.log(this.calcularGanancia(this.jugada, jugador));
         this.ganancia =
           this.ganancia + this.calcularGanancia(this.jugada, jugador);
@@ -97,24 +86,24 @@ export class DeluxeCachinEasyWin extends Juego {
       // Abandona, pierde todo
       if (opcion === "salir") {
         if (this.saldoInicial > jugador.obtenerSaldo()) {
-          this.mostrarResultado("derrota", jugador);
+          this.mostrarResultados("derrota", jugador);
           await this.esperar(3);
           break;
         } else {
-          this.mostrarResultado("victoria", jugador);
+          this.mostrarResultados("victoria", jugador);
           await this.esperar(3);
           break;
         }
       }
       if (opcion === "apuesta") {
-        this.montoApostado += await this.pedirMonto(jugador);
-        if (this.montoApostado === 0) {
+        this.apuesta = await this.pedirApuesta(jugador);
+        if (this.apuesta === 0) {
           continue;
         }
-        await this.interfaceCachin(jugador);
+        await this.interfaceTragamonedas(jugador);
       }
       if (jugador.obtenerSaldo() < 100) {
-        this.mostrarResultado("derrota", jugador);
+        this.mostrarResultados("derrota", jugador);
       }
 
       if (opcion === "instrucciones") {
@@ -132,113 +121,9 @@ export class DeluxeCachinEasyWin extends Juego {
       }
     }
     return {
-      apuestaTotal: this.montoApostado,
+      apuestaTotal: this.apuesta,
       resultado: resultado,
       ganancia: this.ganancia, // Si no hay ganancia, ganancia será 0
     };
-  }
-
-  public simboloRandom() {
-    let i = Math.floor(Math.random() * this.simbolos.length);
-    return this.simbolos[i];
-  }
-
-  public contarParecidos(tirada: string[]): Record<string, number> {
-    const contador: Record<string, number> = {};
-
-    // Cuenta todas los parecidos de cada símbolo en el array
-    for (const simbolo of tirada) {
-      contador[simbolo] = (contador[simbolo] || 0) + 1;
-    }
-
-    return contador;
-  }
-
-  public contarSimilitudes(tirada: string[]): boolean {
-    const contador = this.contarParecidos(tirada);
-
-    // Verifica si algún símbolo aparece más de una vez
-    for (const simbolo in contador) {
-      if (contador[simbolo] > 1) {
-        return true; // Hay similitudes en el array
-      }
-    }
-
-    return false; // No hay similitudes
-  }
-
-  public calcularGanancia(tirada: any, jugador: Jugador): number {
-    let contador = this.contarParecidos(tirada);
-    let gananciaTotal = 0;
-    for (const simbolo in contador) {
-      if (contador[simbolo] >= 2) {
-        // Verifica que se repita al menos dos veces
-        const valorSimbolo = this.valoresSimbolos[simbolo];
-        gananciaTotal +=
-          (this.montoApostado + valorSimbolo) * contador[simbolo];
-        jugador?.sumarSaldo(gananciaTotal);
-      }
-    }
-    if (gananciaTotal > 0) {
-      console.log("\n😃 Ganaste: ");
-    } else if (gananciaTotal === 0) {
-      console.log("\n😔 No hubo suerte esta vez:");
-    }
-    return gananciaTotal;
-  }
-
-  private async tirada() {
-    this.jugada = [];
-    for (let i: number = 0; i < this.simbolos.length; i++) {
-      let cargarSimbolos = this.simboloRandom();
-      this.jugada.push(cargarSimbolos);
-    }
-    return this.jugada;
-  }
-
-  private async esperar(segundos: number): Promise<void> {
-    return new Promise((resolve) => setTimeout(resolve, segundos * 1000));
-  }
-
-  private async pedirMonto(jugador: Jugador) {
-    const montoApostado = await Menu.pedirNumero(
-      "Ingrese su apuesta [0: Para salir]",
-      (apuesta) => {
-        // Primero valida que sea un numero
-        if (typeof apuesta === "number") {
-          // Si es numero
-          // chequea que lo ingresado no sea menor q la apuesta minima
-          // y distinto de 0
-          if (apuesta < 0) {
-            return "Debes ingresar un número válido";
-          }
-          if (apuesta >= 1 && apuesta < this.apuestaMinima) {
-            return `El monto ingresado (${apuesta}) es inferior al minimo requerido (${this.apuestaMinima})`;
-          }
-          // despues, chequea que no supere la apuesta maxima
-          if (apuesta > this.apuestaMaxima) {
-            return `El monto ingresado (${apuesta}) es superior al maximo permitido (${this.apuestaMaxima})`;
-          }
-          if (apuesta > jugador.obtenerSaldo()) {
-            return "Saldo insuficiente.";
-          }
-          return true;
-        } else {
-          return "Debes ingresar un número válido.";
-        }
-      }
-    );
-    return montoApostado;
-  }
-
-  private async interfaceCachin(jugador: Jugador) {
-    console.clear();
-    console.log("========================================================");
-    console.log("             🎰 Deluxe Cachin Easy Win 🎰              ");
-    console.log("========================================================");
-    console.log(
-      ` Bienvenido, ${jugador.obtenerNombre()}     🤑 Saldo: ${jugador.obtenerSaldo()}             `
-    );
-    console.log("--------------------------------------------------------");
   }
 }
